Now works with recent adf.ly updates
Instead of sending you to the link's destination, the url variable in an adfly page now sends you to /go/{random crap here}, which is a relative path. This broke the local proxy because it didn't check for relative paths. In addition, the new /go/{crap} page noticed that a cookie that was modified after you waited 5 seconds wasn't changed and redirected you to a block page accusing you of haxorz.

Fixes:
* proxy.geturl now clears cookies *after* finding the destination
* proxy.adfly assumes that the url variable contains a relative path

Eventually it would be great for proxy.adfly to return the actual destination instead of /go/{crap}.